feat: Varsayılan kategoriler için yeni API uç noktaları eklendi ve kullanıcı rolü ile yönetim yetkileri entegre edildi.
- Kategori işlemlerini gerçekleştirmek için kategori tıklanabilir hale getirildi.
- Silme işlemi modal içerisine taşındı ve modal üzerine kapatma butonu eklendi.

+++ client/src/services/categoryService.js
@@ -71,11 +71,12 @@ class CategoryService {
   /**
    * @description Adds a new category.
    * @param { Object } categoryData - The data of the category to add.
+   * @param { boolean } useAdminRoute - Whether to use the admin route for the request.
    * @return { Promise<{ success: boolean, data?: any, error?: string }> }
    */
-  async addCategory(categoryData) {
+  async addCategory(categoryData, useAdminRoute = false) {
     try {
-      const response = await api.post(`${API_CONFIG.BASE_URL}${API_CONFIG.ENDPOINTS.CATEGORY.ADD}`, categoryData);
+      const response = await api.post(`${API_CONFIG.BASE_URL}${useAdminRoute ? API_CONFIG.ENDPOINTS.CATEGORY.ADMIN.ADD : API_CONFIG.ENDPOINTS.CATEGORY.ADD}`, categoryData);
 
       const data = response.data;
       if (data.success) {
@@ -93,11 +94,13 @@ class CategoryService {
    * @description Updates an existing category.
    * @param { string } categoryId - The ID of the category to update.,
    * @param { Object } categoryData - The updated data for the category.
+   * @param { boolean } useAdminRoute - Whether to use the admin route for the request.
    * @return { Promise<{ success: boolean, data?: any, error?: string }> } 
    */
-  async updateCategory(categoryId, categoryData) {
+  async updateCategory(categoryId, categoryData, useAdminRoute = false) {
+    console.log(categoryData);
     try {
-      const response = await api.put(`${API_CONFIG.BASE_URL}${API_CONFIG.ENDPOINTS.CATEGORY.EDIT}${categoryId}`, categoryData);
+      const response = await api.put(`${API_CONFIG.BASE_URL}${useAdminRoute ? API_CONFIG.ENDPOINTS.CATEGORY.ADMIN.EDIT : API_CONFIG.ENDPOINTS.CATEGORY.EDIT}${categoryId}`, categoryData);
 
       const data = response.data;
       if (data.success) {
@@ -114,11 +117,12 @@ class CategoryService {
   /**
    * @description Deletes a category.
    * @param { string } categoryId - The ID of the category to delete.
+   * @param { boolean } useAdminRoute - Whether to use the admin route for the request.
    * @return { Promise<{ success: boolean, data?: any, error?: string }> }
    */
-  async deleteCategory(categoryId) {
+  async deleteCategory(categoryId, useAdminRoute = false) {
     try {
-      const response = await api.delete(`${API_CONFIG.BASE_URL}${API_CONFIG.ENDPOINTS.CATEGORY.DELETE}${categoryId}`);
+      const response = await api.delete(`${API_CONFIG.BASE_URL}${useAdminRoute ? API_CONFIG.ENDPOINTS.CATEGORY.ADMIN.DELETE : API_CONFIG.ENDPOINTS.CATEGORY.DELETE}${categoryId}`);
 
       const data = response.data;
       if (data.success) {
